Port handler is truly cross-platform
